+++ src/JavLuv/MainWindowViewModel.cs
@@ -31,8 +31,12 @@ namespace JavLuv
             // Log version
             var currentVersion = SemanticVersion.Current;
             Logger.WriteInfo("Running JavLuv version " + currentVersion);
+            bool upgradeVersion = false;
             if (JavLuv.Settings.Get().LastVersionRun < currentVersion)
+            {
                 Logger.WriteInfo("Upgraded from version " + JavLuv.Settings.Get().LastVersionRun);
+                upgradeVersion = true;
+            }
 
             // Catch any unhandled exceptions and display them
             AppDomain.CurrentDomain.UnhandledException += (sender, args) =>
@@ -75,7 +79,7 @@ namespace JavLuv
             // Set state initially
             ChangeState(null);
 
-            // Check version
+            // Check for new version if conditions are met
             var timeToCheck = new TimeSpan(1, 0, 0, 0); // 1 day interval
             var interval = new TimeSpan(0, 0, 0, 0);
             if (JavLuv.Settings.Get().LastVersionCheckTime != null)
@@ -88,11 +92,9 @@ namespace JavLuv
                 CommandQueue.LongTask().Execute(m_checkVersion);
             }
 
-            // Perform a one-time cleanup of actress image folder to fix previous bugs
-            if (JavLuv.Settings.Get().LastVersionRun < new SemanticVersion(1, 1, 6))
-            {
+            // Perform a cleanup of actress image folder on version upgrade to fix previous bugs
+            if (upgradeVersion)
                 Collection.CleanActressImages();
-            }
         }
 
         #endregion